+++ apm-agent-attach/src/test/java/co/elastic/apm/attach/ElasticApmAttacherTest.java
@@ -11,9 +11,9 @@
  * the Apache License, Version 2.0 (the "License"); you may
  * not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *   http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
  * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
@@ -27,6 +27,11 @@ package co.elastic.apm.attach;
 import org.junit.jupiter.api.Test;
 import wiremock.org.apache.commons.codec.digest.DigestUtils;
 
+import java.io.File;
+import java.io.FileReader;
+import java.util.Map;
+import java.util.Properties;
+
 import static org.assertj.core.api.Assertions.assertThat;
 
 class ElasticApmAttacherTest {
@@ -35,5 +40,16 @@ class ElasticApmAttacherTest {
     void testHash() throws Exception {
         assertThat(ElasticApmAttacher.md5Hash(getClass().getResourceAsStream(ElasticApmAttacher.class.getSimpleName() + ".class")))
             .isEqualTo(DigestUtils.md5Hex(getClass().getResourceAsStream(ElasticApmAttacher.class.getSimpleName() + ".class")));
+    }
+
+    @Test
+    void testCreateTempProperties() throws Exception {
+        File tempProperties = ElasticApmAttacher.createTempProperties(Map.of("foo", "bär"));
+        assertThat(tempProperties).isNotNull();
+        tempProperties.deleteOnExit();
+        Properties properties = new Properties();
+        properties.load(new FileReader(tempProperties));
+        assertThat(properties.get("foo")).isEqualTo("bär");
+
     }
 }