+++ CHANGELOG.md
@@ -18,6 +18,7 @@
  * Error in log when setting [server_urls](https://www.elastic.co/guide/en/apm/agent/java/current/config-reporter.html#config-server-urls) 
  to an empty string - `co.elastic.apm.agent.configuration.ApmServerConfigurationSource - Expected previousException not to be null`
 * Avoid terminating the TCP connection to APM Server when polling for configuration updates (#823)
+* Fixes potential segfault if attaching the agent with long arguments (#865)
  
 # 1.9.0